requirements.txt
# RAG Project — Requirements
# Install with: pip install -r requirements.txt

# ── LangChain core ─────────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0

# ── Vector store ───────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ── Embeddings & re-ranking ────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ── PDF parsing (Marker) ───────────────────────────────────────────────────────
marker-pdf>=1.0.0

# ── File loaders ───────────────────────────────────────────────────────────────
pandas>=2.0.0
docx2txt>=0.8
unstructured>=0.14.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
